OSIS-5056 Move clearclipboard to program_management apps + Fix clear action
